src/requirements.txt
selenium
chromedriver-autoinstaller
gspread
google-auth